Fixed multiple issues where data could stopped being consumed.
If the consumer failed to connect to kafka for any reason, the
ArchivingWorker thread dies without retrying.

There is also another issues with the logic that tries to respawn the
thread if it detects an error while consuming, however this could lead
to data loss due to an issue in Kafka (KAFKA-919).